+++ w7/entity.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <cstdint>
 
-constexpr uint16_t invalid_entity = -1;
+constexpr uint32_t invalid_entity = -1;
 struct Entity
 {
   uint32_t color = 0xff00ffff;
@@ -11,7 +11,7 @@ struct Entity
   float ori = 0.f;
   float thr = 0.f;
   float steer = 0.f;
-  uint16_t eid = invalid_entity;
+  uint32_t eid = invalid_entity;
 };
 
 void simulate_entity(Entity &e, float dt);
